# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A chat service api with websocket.",
        "title": "MediChat API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "MediChat API Support",
            "email": "ducalixte.@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/DuCalixte/MediChat-Users/blob/master/LICENSE"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/signin": {
            "post": {
                "description": "authenticates the user access and provides the user resource with a token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signin"
                ],
                "summary": "Authenticate a user access",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Token": {
                                "type": "object",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            }
        },
        "/api/signup": {
            "post": {
                "description": "authorizes the user access and provides the user resource with a token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "signin"
                ],
                "summary": "Authorize a user to access",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        },
                        "headers": {
                            "Token": {
                                "type": "object",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/channels": {
            "get": {
                "description": "get channels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List channels",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name search by q",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Channel"
                            }
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create a new channel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "Create channel",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Channel"
                        },
                        "headers": {
                            "Token": {
                                "type": "object",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/channels/{id}": {
            "get": {
                "description": "get string by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show information on specific channel",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Channel"
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update by json channel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "channels"
                ],
                "summary": "Update an channel",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Channel"
                        }
                    }
                }
            }
        },
        "/api/v1/userPrefs/{id}": {
            "get": {
                "description": "provides user profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show information on specific user profile",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UserPref ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserPref"
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update by json userPref",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "userPrefs"
                ],
                "summary": "Update an userPref",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "UserPref ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserPref"
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "description": "get all users",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List all Users",
                "parameters": [
                    {
                        "type": "string",
                        "description": "name search by q",
                        "name": "q",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.User"
                            }
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "description": "provides user information with preferences in models.UserPref and channels from models.Channel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Show information on specific user",
                "operationId": "get-string-by-int",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        },
                        "headers": {
                            "Token": {
                                "type": "string",
                                "description": "qwerty"
                            }
                        }
                    }
                }
            }
        },
        "/user/{userId}/channel/{id}": {
            "get": {
                "description": "gets User ID and Channel ID and creates a new client connection",
                "consumes": [
                    "application/json"
                ],
                "summary": "Connects a user on a channel to the websocket circuit",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Channel ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.Channel": {
            "type": "object",
            "properties": {
                "chatBot": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "isPrivate": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "ownerId": {
                    "type": "integer"
                },
                "users": {
                    "description": "WebSocket\t\t string\t\t`json:\"websocket\"  gorm:\"type:varchar(256)\"`",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "channels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Channel"
                    }
                },
                "email": {
                    "type": "string"
                },
                "profile": {
                    "type": "object",
                    "$ref": "#/definitions/models.UserPref"
                }
            }
        },
        "models.UserPref": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "gravatar": {
                    "type": "string"
                },
                "isLocal": {
                    "type": "boolean"
                },
                "lastname": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        }
    }
}